More refactoring: split lara-plugin-api into smaller modules
[#165062314]
