+++ docs/lara-plugin-api/interfaces/ipluginstatepath.md
@@ -23,7 +23,7 @@
 
 **● loadPath**: *`string`*
 
-*Defined in [api/plugins.ts:61](https://github.com/concord-consortium/lara/blob/93f2901a/lara-plugin-api/src/api/plugins.ts#L61)*
+*Defined in [api/plugins.ts:61](https://github.com/concord-consortium/lara/blob/30e7426a/lara-plugin-api/src/api/plugins.ts#L61)*
 
 ___
 <a id="savepath"></a>
@@ -32,7 +32,7 @@ ___
 
 **● savePath**: *`string`*
 
-*Defined in [api/plugins.ts:60](https://github.com/concord-consortium/lara/blob/93f2901a/lara-plugin-api/src/api/plugins.ts#L60)*
+*Defined in [api/plugins.ts:60](https://github.com/concord-consortium/lara/blob/30e7426a/lara-plugin-api/src/api/plugins.ts#L60)*
 
 ___
