+++ docs/lara-plugin-api/interfaces/iruntimecontextexperimentalfeatures.md
@@ -22,7 +22,7 @@
 
 **● clickToPlayId**: *`string`*
 
-*Defined in [api/plugins.ts:56](https://github.com/concord-consortium/lara/blob/93f2901a/lara-plugin-api/src/api/plugins.ts#L56)*
+*Defined in [api/plugins.ts:56](https://github.com/concord-consortium/lara/blob/30e7426a/lara-plugin-api/src/api/plugins.ts#L56)*
 
 DOM id of click to play overlay if enabled.
